+++ tests/src/rules/no-internal-modules.js
@@ -7,6 +7,7 @@ const ruleTester = new RuleTester()
 
 ruleTester.run('no-internal-modules', rule, {
   valid: [
+    // imports
     test({
       code: 'import a from "./plugin2"',
       filename: testFilePath('./internal-modules/plugins/plugin.js'),
@@ -57,9 +58,44 @@ ruleTester.run('no-internal-modules', rule, {
         allow: [ '**/index{.js,}' ],
       } ],
     }),
+    // exports
+    test({
+      code: 'export {a} from "./internal.js"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/index.js'),
+    }),
+    test({
+      code: 'export * from "lodash.get"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/index.js'),
+    }),
+    test({
+      code: 'export {b} from "@org/package"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+    }),
+    test({
+      code: 'export {b} from "../../api/service"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        allow: [ '**/api/*' ],
+      } ],
+    }),
+    test({
+      code: 'export * from "jquery/dist/jquery"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        allow: [ 'jquery/dist/*' ],
+      } ],
+    }),
+    test({
+      code: 'export * from "./app/index.js";\nexport * from "./app/index"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        allow: [ '**/index{.js,}' ],
+      } ],
+    }),
   ],
 
   invalid: [
+    // imports
     test({
       code: 'import "./plugin2/index.js";\nimport "./plugin2/app/index"',
       filename: testFilePath('./internal-modules/plugins/plugin.js'),
@@ -123,6 +159,73 @@ ruleTester.run('no-internal-modules', rule, {
           message: 'Reaching to "debug/node" is not allowed.',
           line: 1,
           column: 17,
+        },
+      ],
+    }),
+    // exports
+    test({
+      code: 'export * from "./plugin2/index.js";\nexport * from "./plugin2/app/index"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        allow: [ '*/index.js' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "./plugin2/app/index" is not allowed.',
+        line: 2,
+        column: 15,
+      } ],
+    }),
+    test({
+      code: 'export * from "./app/index.js"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      errors: [ {
+        message: 'Reaching to "./app/index.js" is not allowed.',
+        line: 1,
+        column: 15,
+      } ],
+    }),
+    test({
+      code: 'export {b} from "./plugin2/internal"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      errors: [ {
+        message: 'Reaching to "./plugin2/internal" is not allowed.',
+        line: 1,
+        column: 17,
+      } ],
+    }),
+    test({
+      code: 'export {a} from "../api/service/index"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        allow: [ '**/internal-modules/*' ],
+      } ],
+      errors: [
+        {
+          message: 'Reaching to "../api/service/index" is not allowed.',
+          line: 1,
+          column: 17,
+        },
+      ],
+    }),
+    test({
+      code: 'export {b} from "@org/package/internal"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      errors: [
+        {
+          message: 'Reaching to "@org/package/internal" is not allowed.',
+          line: 1,
+          column: 17,
+        },
+      ],
+    }),
+    test({
+      code: 'export {get} from "debug/node"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      errors: [
+        {
+          message: 'Reaching to "debug/node" is not allowed.',
+          line: 1,
+          column: 19,
         },
       ],
     }),